Modularize code whilst avoiding loops
